Configure linters to run on all files
By defualt, `pre-commit` is supposed to be running checks against only
the files modified by the triggering commit, but in #10 it doesn't
look like the linters ran on anything.

Rather than figuring out why, just configure the workflow to run
`pre-commit` on all files in the repository.

And also explicitly configure yamllint while we're at it.

x-branch: feature/ci