Fix time specification in certificate
While the value e.g. `360h` is valid input for the `duration` or
`renewBefore` keys in the certificate, the canonical form includes
minutes and seconds as well (`360h0m0s`). The value is transformed
when the manifest is loaded on the cluster, so from ArgoCD's
perspective it never matches.

This commit updates the time specification to match the canonical
format.