Add missing clusterrolebinding cluster-admins-rb
We were missing the clusterrolebinding for the cluster-admins group,
which means right now members of cluster-admins have no privileges.
This commit fixes that.

Also note that in addition to the clusterrolebinding named
`cluster-admins-rb` that is created by these manifests, there are also
clusterolebindings created by openshift named `cluster-admin` and
`cluster-admins`, so be sure you're looking at the right one :).