Use bundles to simplify management of overlays
When we add a cluster-wide feature, this often requires adding
multiple manifests to the cluster-scope overlay (namespaces, roles,
rolebindings, crds, etc). This can complicate management of the
overlay, since it's possible to accidentally omit a required manifest
(and it's possible to accidentally leave a manifest in place when
removing the feature).

Bundles are a way of packaging together a set of related resources so
that they can be enabled/disabled as a group.

x-branch: feature/bundles