Create letsencrypt clusterissuer
We have existing cert-manager issuers in the openshift-ingress
namespace, but now that we are looking at creating certificates in
other namespaces it makes sense to replace this with a ClusterIssuer
resource.

This commit creates the ClusterIssuer; a subsequent pr will remove the
namespaced-issuers and update the appropriate certificate requests.

Part of cci-moc/ops-issues#443