Perform schema validation on pull requests
Use kubeval [1] to validate the output of `kustomize build` for all of
our overlays. This relies on the cci-moc/openshift-schemas repository
[2], which is generated by scraping the OpenShift OpenAPI endpoint and
passing that through openapi2jsonchema [3].

With our current CI, we check only for proper YAML syntax, but this
won't find typos, missing keys, or other problems (for example,
see #72). Performing schema validation on pull requests will detect
these sort of problems earlier.

This does mean that when adding new CRDs to the cluster we will also
need to update the openshift-schemas repository.

[1]: https://github.com/instrumenta/kubeval
[2]: https://github.com/CCI-MOC/openshift-schemas
[3]: https://github.com/instrumenta/openapi2jsonschema