[netbox] Work around kustomize version skew
Kustomize 4.4.0 [1], which I am running locally, has much improved support
for YAML anchors and aliases. Our argocd environment is running 4.3.0,
which was failing to render the manifests in this application
correctly.

I've rewritten the manifests to work correctly with both versions, but
once operate-first/SRE#421 is resolved we can revert these changes.

[1]: https://github.com/kubernetes-sigs/kustomize/releases/tag/kustomize%2Fv4.4.0