Use default strategy for Netbox deployment
The netbox `Deployment` was configured to use the `Recreate` strategy,
which means that a replacement pod can't start until the existing
pod has terminated. This is unnecessary and slows down the process of
rolling out configuration changes.

This commit removes the explicit `strategy.type` setting from the
`Deployment`.

x-branch: feature/no-recreate-deploy